Add `use experimental :pack`
Message.pm6 is using experimental pack/unpack rakudo feature which
needs to be enabled with a pragma. Rakudo versions ≤ 2018.05 used to
give a run time error on first use, but starting with 2018.06 it will
be a compile time error. I think tests were never hitting the lines
with `pack` which is why tests were green previously.